agent_adapter/nanobot-channel/pyproject.toml
[project]
name = "nanobot-channel-agent-club"
version = "1.0.0"
description = "Nanobot channel plugin for Agent Club IM"
requires-python = ">=3.10"
dependencies = [
    "python-socketio[asyncio_client]>=5.10",
    "aiohttp>=3.9",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.entry-points."nanobot.channels"]
agent_club = "nanobot_channel_agent_club:AgentClubChannel"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
